requirements.txt
requests
py-cord[voice]
yt-dlp
python-dotenv
